Strategy #1: restore state using v-model as elements are rendered
Pass the above state down as a prop called `state`.

When rendering each **Select** instance, set its *input* value to the corresponding property of the `state` (using `v-model` here makes all `state` properties reactive, meaning they will be updated if the inputs are changed)

This allow state to be restored as the UI component is rendered.